Moves Importers::Base to importers/base.rb
